Keep Follow input schemas with Workflows
