+++ apps/worker/package.json
@@ -19,7 +19,8 @@
     "@temporalio/workflow": "^1.22.0",
     "drizzle-orm": "1.0.0-beta.22",
     "ts-pattern": "^5.9.0",
-    "tsx": "^4.22.4"
+    "tsx": "^4.22.4",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
     "@fedify/fedify": "^2.3.0",
