+++ packages/core/package.json
@@ -51,7 +51,7 @@
     "test:services": "node --import tsx --import ./temporal/test-client.ts --test --test-concurrency=1 'services/*.test.ts'",
     "test:services:database": "cd ../.. && node scripts/test-db.mjs run -- pnpm --filter @kosmo/core test:services:isolated",
     "test:services:isolated": "cd ../.. && pnpm db:test:push && pnpm --filter @kosmo/core test:services",
-    "test:unit": "node --import tsx --test activitypub-note-content.test.ts 'post-content/**/*.test.ts' temporal/client.test.ts temporal/follow-command.test.ts validation/profile-tag.test.ts validation/profile-follow.test.ts db/index.test.ts"
+    "test:unit": "node --import tsx --test activitypub-note-content.test.ts 'post-content/**/*.test.ts' temporal/client.test.ts temporal/follow-command.test.ts validation/profile-tag.test.ts db/index.test.ts"
   },
   "devDependencies": {
     "@types/jsdom": "^28.0.3",
