+++ packages/core/validation/index.ts
@@ -1,6 +1,5 @@
 export * from './feedback';
 export * from './post';
 export * from './profile';
-export * from './profile-follow';
 export * from './profile-tag';
 export * from './reaction';
